--- App/views/index.html
@@ -1,54 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8" />
-  <title>Twitter Client</title>
-  <link rel="stylesheet" href="/style.css" />
-</head>
-<body>
-  <div class="container">
-    <h1>🕊️ Twitter Client</h1>
-
-    <form id="tweetForm">
-      <textarea name="tweetText" placeholder="Write your tweet here..." required></textarea>
-      <button type="submit">Post Tweet</button>
-    </form>
-
-    <h2>Recent Tweets</h2>
-    <button onclick="loadTimeline()">Refresh Timeline</button>
-    <ul id="tweetList"></ul>
-  </div>
-
-  <script>
-    const tweetForm = document.getElementById('tweetForm');
-    const tweetList = document.getElementById('tweetList');
-
-    tweetForm.addEventListener('submit', async (e) => {
-      e.preventDefault();
-      const tweetText = e.target.tweetText.value;
-
-      const response = await fetch('/twitter/tweet', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ text: tweetText })
-      });
-
-      const result = await response.json();
-      alert(result.message);
-      e.target.reset();
-    });
-
-    async function loadTimeline() {
-      const response = await fetch('/twitter/timeline');
-      const data = await response.json();
-
-      tweetList.innerHTML = '';
-      data.data?.forEach(tweet => {
-        const li = document.createElement('li');
-        li.textContent = tweet.text;
-        tweetList.appendChild(li);
-      });
-    }
-  </script>
-</body>
-</html>
