+++ routes/twitter.js
@@ -3,7 +3,7 @@ const router = express.Router();
 const { postTweet, getRecentTweets } = require('../controllers/twitterController');
 
 // Route to post a tweet
-router.post('/tweet', postTweet);
+//router.post('/tweet', postTweet);
 
 // Route to get recent tweets
 router.get('/timeline', getRecentTweets);
